+++ mayan/apps/sources/literals.py
@@ -14,6 +14,8 @@ DEFAULT_SOURCES_CACHE_STORAGE_BACKEND_ARGUMENTS = {
 
 DEFAULT_SOURCES_LOCK_EXPIRE = 600
 
+ERROR_LOG_DOMAIN_NAME = 'sources'
+
 SOURCE_ACTION_EXECUTE_TASK_PATH = 'mayan.apps.sources.tasks.task_source_backend_action_execute'
 
 STORAGE_NAME_SOURCE_CACHE_FOLDER = 'sources__source_cache'